Added high quality images and new structure
