+++ lib/search.js
@@ -33,4 +33,28 @@ function simpleSearch (term) {
   }).catch(ifError)
 }
 
-module.exports = {search, simpleSearch}
+function searchActor(keyword){
+  return request(`https://www.imdb.com/find?ref_=nv_sr_fn&q=${keyword}&s=nm`).then((data) => {
+    const $ = cheerio.load(data);
+    let result = [];
+    result.push({
+      searchQuery: keyword
+    });
+    $('tr').each((index, element) => {
+      try{
+        var image = $(element).find('td.primary_photo > a > img').attr('src');
+        var name = $(element).find('td.result_text > a').text();
+        image = image.replace('"', "").replace(/\\g/, "").split("@._")[0] += '@._V1_UX1024_CR1024,1024,0,0_AL_.jpg';
+        result.push({
+          actorImage: image,
+          actorName: name
+        });
+      }catch(e){
+        console.log("An error occurred when pushing: " + e);
+      }
+    });
+    return result;
+  }).catch(ifError)
+}
+
+module.exports = {search, simpleSearch, searchActor}
